test: "CCYYMMDD" is not a valid XSD date
XSD-date cannot be written in the form "CCYYMMDD", this patch adds a
test to ensure it.

linkml-runtime has been relying on Python `fromisoformat` [1] format to
parse and, implicitly validate, XSD-date. But from Python 3.11 upwards
`fromisoformat` also accepts "CCYYMMDD" as a valid date, breaking
XSD-date compatibility.

This is a regression test to avoid this issue in the future.

[1]: https://docs.python.org/3/library/datetime.html#datetime.date.fromisoformat

Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>